feat(interactable-players): interactable players with interactions for current game play
